Add greetd config
Add files (but don't auto-install for now, as this overwrites dpkg
provided files) so that I remember how I've installed greetd and
wlgreet in the past.  In practice, I'll diff these and then copy them
by hand.

Remove wlgreet from the packages to install, as I need at least
version 1.5 to avoid a bug when launched from greetd and which is not
yet available from dpkg.  So I currently must install from source.

https://git.sr.ht/~kennylevinsen/wlgreet